s3/librpc: smbXsrv.idl: updates for Persistent Handles
Changes in smbXsrv_open_global0:
- persistent
- file_id
- name_hash
- app_instance_vhigh
- app_instance_vlow

The persistent flag is the primary means of directing all FSA and SMB layer
behaviour for Persistent Handles.

The file_id is used when scanning the smbXsrv_open_global.tdb for stale opens
and scheduling cleanup via the scavenger: scavenger uses the file-id from the
open to fetch the file record from locking.tdb.

name_hash will be used as part of implementing delete-on-close token cleanup via
the scavenger.

app_instance_vhigh and app_instance_vlow are part of the app_instance_id related
feature.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Anoop C S <anoopcs@samba.org>